fix bugs

+++ Employee.cpp
@@ -1,6 +1,6 @@
 #include "Employee.hpp"
 
-Employee::Employee(std::string n, std::string s, double sal):
+Employee::Employee(std::string n, std::string s, int sal):
     Record(n, s), salary(sal) {}
 
 double Employee::getSalary() const { return salary; }
